% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_edge_intensity.R
\name{calculate_edge_intensity}
\alias{calculate_edge_intensity}
\title{Calculate Edge Intensity Difference Between Two Images}
\usage{
calculate_edge_intensity(img_df1, img_df2, edge_width = 150)
}
\arguments{
\item{img_df1}{`data.frame`. First image data frame with columns `x`, `y`,
`R`, `G`, `B` created by `initialize_image()`.}

\item{img_df2}{`data.frame`. Second image data frame with columns `x`, `y`,
`R`, `G`, `B` created by `initialize_image()`.}

\item{edge_width}{`numeric(1)`. Width of edge region in pixels from the
outer border. Default is 150.}
}
\value{
A `list` containing:
  \itemize{
    \item `mean_intensity1`: Mean intensity of edge pixels in first image
    \item `mean_intensity2`: Mean intensity of edge pixels in second image
    \item `difference`: Difference between the two mean intensities
  }
}
\description{
Compares the mean pixel intensities in the edge regions of two images.
Extracts pixels within a specified distance from the image borders and
calculates the difference in their mean grayscale intensities. Useful
for background normalization and image comparison tasks.
}
