% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_edge_intensity.R
\name{calculate_edge_intensity_imager}
\alias{calculate_edge_intensity_imager}
\title{Calculate Edge Intensity Difference Between Two Images}
\usage{
calculate_edge_intensity_imager(img1, img2, edge_width = 150)
}
\arguments{
\item{img1}{`cimg`. First image object from the imager package.}

\item{img2}{`cimg`. Second image object from the imager package.}

\item{edge_width}{`numeric(1)`. Width of edge region in pixels from the
outer border. Default is 150.}
}
\value{
A `list` containing:
  \itemize{
    \item `mean_intensity1`: Mean intensity of edge pixels in first image
    \item `mean_intensity2`: Mean intensity of edge pixels in second image
    \item `difference`: Difference between the two mean intensities
  }
}
\description{
Compares the mean pixel intensities in the edge regions of two images.
Extracts pixels within a specified distance from the image borders and
calculates the difference in their mean grayscale intensities. Useful
for background normalization and image comparison tasks.
}
